refactor: move Difference to internal
